+++ imports/layout/application-layout.js
@@ -21,6 +21,7 @@ Template.ApplicationLayout.events({
   },
   "focus #search-form input[name=search]"(e, t) {
     e.preventDefault();
+    $(e.target).blur();
     Modal.show("SearchModal", null, {
       backdrop: 'static',
       keyboard: false