+++ imports/startup/both/index.js
@@ -1,2 +1,3 @@
 import './avatar';
 import './markdown';
+import './lodash';