+++ imports/startup/both/lodash.js
@@ -1,0 +1,3 @@
+import lodash from 'lodash';
+
+_ = lodash;